Splitting realis into regular API and admin API files.
